requirements.txt
python-dotenv==1.1.1
aiogram==3.21.0
pydantic==2.11.7
